pyproject.toml
[project]
name = "metropolitan"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "discord-py>=2.7.1",
    "python-dotenv>=1.2.2",
]
